Move harvest-notifier to harvest_notifier
