+++ config.ru
@@ -4,6 +4,6 @@ lib_dir = File.expand_path("lib", File.dirname(__FILE__))
 $LOAD_PATH << lib_dir unless $LOAD_PATH.include?(lib_dir)
 
 require "harvest_notifier"
-require "harvest-notifier/slack_handler"
+require "harvest_notifier/slack_handler"
 
 run HarvestNotifier::SlackHandler.new
