+++ lib/harvest_notifier/base.rb
@@ -3,10 +3,10 @@
 require "active_support/core_ext/date/calculations"
 require "active_support/core_ext/date_and_time/calculations"
 
-require "harvest-notifier/report"
-require "harvest-notifier/notification"
-require "harvest-notifier/harvest"
-require "harvest-notifier/slack"
+require "harvest_notifier/report"
+require "harvest_notifier/notification"
+require "harvest_notifier/harvest"
+require "harvest_notifier/slack"
 
 module HarvestNotifier
   class Base
