+++ lib/harvest_notifier/harvest.rb
