+++ lib/harvest_notifier/notification.rb
@@ -3,9 +3,9 @@
 require "active_support/core_ext/module/delegation"
 require "active_support/core_ext/string/inflections"
 
-require "harvest-notifier/templates/daily_report"
-require "harvest-notifier/templates/weekly_report"
-require "harvest-notifier/templates/congratulation"
+require "harvest_notifier/templates/daily_report"
+require "harvest_notifier/templates/weekly_report"
+require "harvest_notifier/templates/congratulation"
 
 module HarvestNotifier
   class Notification
