+++ lib/harvest_notifier/rollbar.rb
