+++ lib/harvest_notifier/slack.rb
