+++ lib/harvest_notifier/templates/base.rb
